ENH: add acq_time_precise after acq_time to _scans
Unfortunately BIDS ATM does not instruct possibility to add .microseconds
to acq_time within _scans files, explicitly stops on "seconds".  So to not
break existing tools supporting BIDS and stay compliant, to introduce sub-second
precision we are adding an ad-hoc field (since allowed).

Note: in this commit I have not taken advantage of prior commit which
introduced microseconds keyword argument to get_datetime - to not waste
cpu cycles on reparsing and re-conversion.